Moved contextual menus into components folder
