+++ static/js/publisher/release/components/contextualMenu.js
@@ -44,23 +44,23 @@ export default class ContextualMenu extends Component {
   }
 
   renderIcon() {
-    return this.props.icon || <i className="p-icon--contextual-menu" />;
+    return this.props.label || <i className="p-icon--contextual-menu" />;
   }
 
   render() {
-    const { isDisabled, position } = this.props;
+    const { appearance, isDisabled, position } = this.props;
     const menuClass = "p-contextual-menu" + (position ? `--${position}` : "");
-    const appearance = this.props.appearance || "neutral";
+    const buttonClass = `p-button${appearance ? `--${appearance}` : ""}`;
     const className = [
-      "p-promote-button p-icon-button",
-      `p-button--${appearance}`,
+      "p-promote-button",
+      buttonClass,
       menuClass,
       isDisabled ? "is--disabled" : "",
       this.props.className || ""
     ].join(" ");
 
     return (
-      <span
+      <button
         className={className}
         onClick={isDisabled ? null : this.dropdownButtonClick.bind(this)}
       >
@@ -68,14 +68,14 @@ export default class ContextualMenu extends Component {
         <span className="p-contextual-menu__dropdown" aria-hidden="true">
           {this.props.children}
         </span>
-      </span>
+      </button>
     );
   }
 }
 
 ContextualMenu.propTypes = {
   isDisabled: PropTypes.bool,
-  icon: PropTypes.oneOfType([PropTypes.node, PropTypes.string]),
+  label: PropTypes.oneOfType([PropTypes.node, PropTypes.string]),
   children: PropTypes.node,
   className: PropTypes.string,
   position: PropTypes.oneOf(["left", "center"]), // right is by default
